Sovellus valmis ja Azure konfigurointi tehty

+++ app/views.py
@@ -6,7 +6,6 @@ from django.contrib.auth import authenticate, login, logout
 def loginview(request):
     return render (request, "loginpage.html")
 
-
 # Login action
 def login_action(request):
     user = request.POST['username']
@@ -44,33 +43,48 @@ def productlistview(request):
 
 
 def addproduct(request):
-    a = request.POST['productname']
-    b = request.POST['packagesize']
-    c = request.POST['unitprice']
-    d = request.POST['unitsinstock']
-    e = request.POST['supplier']
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        a = request.POST['productname']
+        b = request.POST['packagesize']
+        c = request.POST['unitprice']
+        d = request.POST['unitsinstock']
+        e = request.POST['supplier']
     
     Product(productname = a, packagesize = b, unitprice = c, unitsinstock = d, supplier = Supplier.objects.get(id = e)).save()
     return redirect(request.META['HTTP_REFERER'])
 
 def confirmdeleteproduct(request, id):
-    product = Product.objects.get(id = id)
-    context = {'product': product}
-    return render (request,"confirmdelprod.html",context)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        product = Product.objects.get(id = id)
+        context = {'product': product}
+        return render (request,"confirmdelprod.html",context)
 
 
 def deleteproduct(request, id):
-    Product.objects.get(id = id).delete()
-    return redirect(productlistview)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        Product.objects.get(id = id).delete()
+        return redirect(productlistview)
 
 
 def edit_product_get(request, id):
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
         product = Product.objects.get(id = id)
         context = {'product': product}
         return render (request,"edit_product.html",context)
 
 
 def edit_product_post(request, id):
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
         item = Product.objects.get(id = id)
         item.unitprice = request.POST['unitprice']
         item.unitsinstock = request.POST['unitsinstock']
@@ -79,10 +93,13 @@ def edit_product_post(request, id):
 
 
 def products_filtered(request, id):
-    productlist = Product.objects.all()
-    filteredproducts = productlist.filter(supplier = id)
-    context = {'products': filteredproducts}
-    return render (request,"productlist.html",context)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        productlist = Product.objects.all()
+        filteredproducts = productlist.filter(supplier = id)
+        context = {'products': filteredproducts}
+        return render (request,"productlist.html",context)
 
 
 
@@ -97,21 +114,27 @@ def supplierlistview(request):
 
 
 def addsupplier(request):
-    a = request.POST['companyname']
-    b = request.POST['contactname']
-    c = request.POST['address']
-    d = request.POST['phone']
-    e = request.POST['email']
-    f = request.POST['country']
-    Supplier(companyname = a, contactname = b, address = c, phone = d, email = e, country = f).save()
-    return redirect(request.META['HTTP_REFERER'])
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        a = request.POST['companyname']
+        b = request.POST['contactname']
+        c = request.POST['address']
+        d = request.POST['phone']
+        e = request.POST['email']
+        f = request.POST['country']
+        Supplier(companyname = a, contactname = b, address = c, phone = d, email = e, country = f).save()
+        return redirect(request.META['HTTP_REFERER'])
 
 
 def searchsuppliers(request):
-    search = request.POST['search']
-    filtered = Supplier.objects.filter(companyname__icontains=search)
-    context = {'suppliers': filtered}
-    return render (request,"supplierlist.html",context)
+    if not request.user.is_authenticated:
+        return render(request, 'loginpage.html')
+    else:
+        search = request.POST['search']
+        filtered = Supplier.objects.filter(companyname__icontains=search)
+        context = {'suppliers': filtered}
+        return render (request,"supplierlist.html",context)
